Update IsEnemyElite.lua
This major change should now essentially ensure all elites slain are accounted for regardless of targeting. It works by scanning all enemy units on the threat table for all party and raid members and counts them if they are killed. Allowing for situations where no one is targeting the npc.