Update SetTargetTooltipDisplay.lua
This seems ideal. I did a LOT of testing on this with different approaches (setting text to empty, nil, hiding, etc.) and all of them caused issues or flickering. This approach makes the level line font size 1, and makes it transparent. Ultimately removing it and it seems bulletproof for me.