feat(CanGetSoulshard): added a new function to check if player can get soulshards
the new file checks if the player is a warlock class and checks if the target will give the player experience points. If both conditions are met an icon is shown above the `CustomResourceBar`. This can be changed. Very important notice: I have not updated the `UltraHardcore.toc` import because it might cause merge conflicts. The file has to be added to the `.toc` file after implementation.

Massive shoutout to @tulhur for helping break down the xp levels and also for Missed Vixen and Munkee on the Ultra Discord for suggesting this feature