feat(soulshard): this enables the missing option to enable or disable the feature
I added additional values to be able to set the `ShowSoulShard Indicator` for players in `SettingsOptionsTab.lua`, `LoadDBData.lua` and `PresetSections.lua`. I still have not added anything to the `UltraHardcore.toc` because I was afraid this PR might run into conflicts. Have a very close look at the comments. I kept in the class check (commented it out) and also added some context comments for the indicator icon. All is up fo debate and can be changed.

Thanks again for @tulhur for pre-implementation looking through the code and reminding me to check for the drain soul spell, rather than having a class check. Testing this will only be available for warlocks above level 10.