feat(soulshard): updated the function logic
I extracted the `PlayerKnowsDrainSoul` into a function to make the `CanGainXPFromTarget` a function that could be put into its separate file. Also adjusted the level difference referencing @tulhur's mention from the wowwiki. I'm unsure if these are the correct values, they definitely have to be tested in a real scenario. We can always come back to adjust the values, because I also extracted those into its separate function.

Yet again thanks to @tulhur for doing the testing and QA in this whole process, while not even implemented yet. We're looking at it from many angles.